Instead of going directly through SBL Bridge, we now bridge the call to SBL Bridge through Storage
