+++ src/Altinn.Correspondence.Core/Services/IAltinnStorageService.cs
@@ -1,7 +1,5 @@
-using Altinn.Correspondence.Core.Models.Entities;
-
 namespace Altinn.Correspondence.Core.Services;
-public interface IAltinnSblBridgeService
+public interface IAltinnStorageService
 {
     Task<bool> AddPartyToSblBridge(int partyId, CancellationToken cancellationToken);
