+++ src/Altinn.Correspondence.Integrations/Altinn/Storage/AltinnStorageDevService.cs
@@ -1,0 +1,16 @@
+﻿using Altinn.Correspondence.Core.Services;
+
+namespace Altinn.Correspondence.Integrations.Altinn.Storage
+{
+    public class AltinnStorageDevService : IAltinnStorageService
+    {
+        public async Task<bool> AddPartyToSblBridge(int partyId, CancellationToken cancellationToken = default)
+        {
+            if (partyId <= 0)
+            {
+                return false;
+            }
+            return true;
+        }
+    }
+}
